+++ CHANGELOG.md
@@ -15,12 +15,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Minor fix on authorized types error message.
 - Fix tests according to new contextOps.
+- Fix agents and generators ids for bulk tests.
 
 ### Updated ###
 
 - Update code to take into account last changes on the main project.
 - Bump packages and update code accordingly.
 - Reduce bundle size.
+- Remove support of node 8.
 
 ## [1.19.1](https://github.com/craft-ai/craft-ai-client-js/compare/v1.19.0...v1.19.1) - 2020-05-26 ##